# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.CreateOfferRequest:
    properties:
      application_id:
        type: integer
      benefits:
        type: string
      letter_content:
        type: string
      location:
        type: string
      position:
        type: string
      salary:
        type: integer
      start_date:
        type: string
    required:
    - application_id
    - benefits
    - letter_content
    - location
    - position
    - salary
    - start_date
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ApproverStatus:
    properties:
      approver_id:
        type: integer
      comment:
        type: string
      status:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ListOffersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.OfferItem'
        type: array
      total:
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.OfferDetail:
    properties:
      approvers:
        items:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ApproverStatus'
        type: array
      offer_id:
        type: integer
      status:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.OfferItem:
    properties:
      application_id:
        type: integer
      location:
        type: string
      offer_id:
        type: integer
      position:
        type: string
      salary:
        type: integer
      start_date:
        type: string
      status:
        type: string
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.SendOfferEmailResponse:
    properties:
      email_history_id:
        type: integer
      sent_at:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_request.CreateGridPreference:
    properties:
      config:
        additionalProperties: true
        description: 설정 데이터
        type: object
      grid_name:
        description: 그리드명
        type: string
    required:
    - config
    - grid_name
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_request.UpdateGridPreference:
    properties:
      config:
        additionalProperties: true
        description: 설정 데이터
        type: object
    required:
    - config
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_response.GridPreference:
    properties:
      config:
        additionalProperties: true
        description: 설정 데이터
        type: object
      created_at:
        description: 생성일
        type: string
      created_by:
        description: 수정자
        type: string
      grid_name:
        description: 그리드명
        type: string
      preference_id:
        description: 설정 ID
        type: integer
      updated_at:
        description: 수정일시 (NULL 허용)
        type: string
      updated_by:
        description: 수정자   (NULL 허용)
        type: string
      user_id:
        description: 유저 ID
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.CreateEmailHistory:
    properties:
      applicant_id:
        type: integer
      body:
        type: string
      offer_id:
        type: integer
      template_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    required:
    - body
    - template_id
    - title
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.CreateEmailHistoryResponse:
    properties:
      email_id:
        type: integer
      sent_at:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailHistoryDetail:
    properties:
      body:
        type: string
      email_id:
        type: integer
      sent_at:
        type: string
      title:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailHistoryItem:
    properties:
      email_id:
        type: integer
      sent_at:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: test@example.com
    name: 홍정완
  description: 채용 프로세스 엔진 백엔드 API 문서
  title: Recruit Process Engine API
  version: "1.0"
paths:
...
  /email-history:
    get:
      description: 이메일 발송 이력 목록과 페이징 정보를 반환한다.
      parameters:
      - description: 지원자 식별자
        in: query
        name: applicant_id
        type: integer
      - description: 오퍼 식별자
        in: query
        name: offer_id
        type: integer
      - description: 페이지 번호
        in: query
        name: page
        type: integer
      - description: 페이지당 목록 수
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailHistoryItem'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 이메일 발송 이력 목록 조회
      tags:
      - EmailHistory
    post:
      consumes:
      - application/json
      description: 이메일 발송 이력 레코드를 생성한다.
      parameters:
      - description: 이메일 발송 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.CreateEmailHistory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.CreateEmailHistoryResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 이메일 발송 이력 생성
      tags:
      - EmailHistory
  /email-history/{email_id}:
    get:
      description: 특정 이메일 발송 이력의 상세 정보를 반환한다.
      parameters:
      - description: 이메일 이력 식별자
        in: path
        name: email_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailHistoryDetail'
        "404":
          description: Not Found
      summary: 이메일 발송 이력 상세 조회
      tags:
      - EmailHistory
...
  /offers:
    get:
      description: 오퍼 목록과 총 개수를 조회한다.
      parameters:
      - description: 오퍼 상태 필터
        in: query
        name: status
        type: string
      - description: 페이지 번호
        in: query
        name: page
        type: integer
      - description: 페이지당 항목 수
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ListOffersResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 오퍼 목록 조회
      tags:
      - Offers
    post:
      description: 새로운 오퍼를 생성한다.
      parameters:
      - description: 오퍼 생성 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.CreateOfferRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ListOffersResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 오퍼 생성
      tags:
      - Offers
  /offers/{offer_id}:
    get:
      description: 오퍼의 상세 정보와 승인자 목록을 조회한다.
      parameters:
      - description: 오퍼 식별자
        in: path
        name: offer_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.OfferDetail'
        "404":
          description: Not Found
      summary: 오퍼 상세 조회
      tags:
      - Offers
...
  /offers/{offer_id}/send:
    post:
      description: 오퍼 레터 이메일을 전송하고, 발송 이력을 기록하여 반환한다.
      parameters:
      - description: 오퍼 식별자
        in: path
        name: offer_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.SendOfferEmailResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 오퍼 레터 이메일 전송
      tags:
      - Offers
  /users/grid-preferences:
    get:
      description: 현재 로그인된 사용자의 모든 그리드 설정을 반환한다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_response.GridPreference'
            type: array
        "500":
          description: Internal Server Error
      summary: 그리드 설정 조회
      tags:
      - GridPreferences
    post:
      consumes:
      - application/json
      description: 새로운 그리드 설정을 생성한다.
      parameters:
      - description: 생성할 그리드 설정
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_request.CreateGridPreference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_response.GridPreference'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 그리드 설정 생성
      tags:
      - GridPreferences
  /users/grid-preferences/{preference_id}:
    delete:
      description: 선택된 그리드 설정을 삭제한다.
      parameters:
      - description: 설정 식별자
        in: path
        name: preference_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 그리드 설정 삭제
      tags:
      - GridPreferences
    put:
      consumes:
      - application/json
      description: 그리드 설정(config)을 업데이트한다.
      parameters:
      - description: 설정 식별자
        in: path
        name: preference_id
        required: true
        type: integer
      - description: 업데이트할 설정 내용
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_request.UpdateGridPreference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_preference_dto_response.GridPreference'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 그리드 설정 수정
      tags:
      - GridPreferences
swagger: "2.0"